requirements.txt
ultralytics==8.3.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.10.0.84
numpy==1.26.4
deep-sort-realtime==1.3.2
filterpy==1.4.5
scipy==1.10.1
scikit-image==0.21.0
PyYAML==6.0.2
flask==3.0.3
